Export shared getFormatter() method for grunt-nsp module
