+++ lib/commands/check.js
@@ -3,20 +3,8 @@ var usage = require('../../lib/utils/usage.js')('check.txt');
 var Check = require('../check.js');
 var Formatters = require('../formatters');
 var Path = require('path');
+var getFormatter = require('../../lib/index').getFormatter;
 
-var getFormatter = function (name) {
-
-  if (Object.keys(Formatters).indexOf(name) !== -1) {
-    return Formatters[name];
-  }
-
-  try {
-    return require('nsp-formatter-' + name);
-  }
-  catch (e) {}
-
-  return Formatters.default;
-};
 
 var onCommand = function (args) {
